update course model and amount field in coupon,credit,transaction

+++ ParentProject/BaseModule/src/main/java/BaseModule/eduDAO/CouponDao.java
@@ -119,7 +119,7 @@ public class CouponDao {
 
 	private static Coupon createCouponByResultSet(ResultSet rs) throws SQLException {
 		return new Coupon(rs.getLong("couponId"), rs.getInt("bookingId"), rs.getInt("transactionId"), rs.getInt("userId"),
-				rs.getDouble("amount"), DateUtility.DateToCalendar(rs.getTimestamp("creationTime")), 
+				rs.getInt("amount"), DateUtility.DateToCalendar(rs.getTimestamp("creationTime")), 
 				DateUtility.DateToCalendar(rs.getTimestamp("expireTime")),CouponStatus.fromInt(rs.getInt("status")));
 	}
